refactor(ui): Move MultiSection, MultiSectionNavBar to common
