+++ apps/clearance_ui/src/components/common/MultiSectionNavBar.tsx
