+++ apps/clearance_ui/src/pages/admin/users/index.tsx
@@ -7,7 +7,7 @@ import axios from "axios";
 import { Loader2 } from "lucide-react";
 import { useRouter } from "next/router";
 import { userHooks } from "@/hooks/zodiosHooks";
-import MultiSection from "@/components/MultiSection";
+import MultiSection from "@/components/common/MultiSection";
 import RegisterUser from "@/components/user_management/RegisterUser";
 
 export default function UserManagement() {
